bugfix: remove closeModal() call after delete task func to squash memory leak in task card drop down
