+++ client/components/TaskCardDropDown.js
@@ -2,11 +2,7 @@ import React, {useState, useContext} from 'react'
 import Modal from './Modal'
 import {ColumnContext} from '../context/columnContext'
 import DeleteTaskModal from './DeleteTaskModal'
-import socket from '../socket'
 import styles from './css/TaskCardDropDown.css'
-
-import axios from 'axios'
-import {ProjectContext} from '../context/projectContext'
 
 // fields are actions that user can take from dropdown menu
 const fields = [
@@ -31,30 +27,6 @@ const TaskCardDropDown = ({task}) => {
     isSingleTaskVisible,
     setSingleTaskVisible,
   } = useContext(ColumnContext)
-
-  // grab tasks, setTasks from column context
-  const {columns, setColumns, tasks, setTasks} = useContext(ProjectContext)
-
-  const deleteTask = async () => {
-    try {
-      await axios.delete(`/api/tasks/${task.id}`)
-
-      const {data} = await axios.get(`/api/columns/${task.columnId}`)
-
-      // the next two calls look inefficient but are absolutely necessary
-      // to avoid memory leaks, we set columns before resetting tasks
-
-      setColumns(
-        columns.map((column) => (column.id === data.id ? data : column))
-      )
-
-      setTasks(tasks.filter((currTask) => currTask.id !== task.id))
-    } catch (err) {
-      console.error(err)
-    }
-    socket.emit('delete-task', {ignore: socket.id})
-    closeModal()
-  }
 
   // handle drop down selections
   const handleSelectOption = (option) => {
@@ -85,7 +57,7 @@ const TaskCardDropDown = ({task}) => {
       <div className={styles.arrowDown}></div>
       {activeField === 'Delete' && (
         <Modal>
-          <DeleteTaskModal deleteTask={deleteTask} closeModal={closeModal} />
+          <DeleteTaskModal task={task} closeModal={closeModal} />
         </Modal>
       )}
     </div>
